Tweak how activation commands are generated
